Moved code from InstPrinter -> Disassembler
